requirements.txt
fastapi
uvicorn
openai-whisper
torch
python-multipart
requests
python-dotenv
